Renamed class to `UserFactory::WithPassword`.
RSpec: 998 examples, 0 failures, 1 pending
RuboCop: 288 files inspected, no offenses detected
Coverage: 3086 / 3100 LOC (99.55%) covered.
