+++ spec/controllers/sessions_controller_spec.rb
@@ -39,7 +39,7 @@ end # shared_examples 'invalid login credentials'
 describe SessionsController do
   let(:registered_user) do
     user_attribs = FactoryGirl.attributes_for(:user, :saved_user)
-    user = UserPasswordEntityFactory.create user_attribs, 'password'
+    user = UserFactory::WithPassword.create user_attribs, 'password'
     UserRepository.new.add user
     user
   end
